Move query code out of the plugin file
